Stable?

+++ packs/BidInfo.java
@@ -1,20 +1,20 @@
 package packs; // Ensure this matches the package name of your other classes, if they are intended to be together.
 
 public class BidInfo {
-    private int bidAmount;
+    private Double bidAmount;
     private int buyerId;
 
-    public BidInfo(int buyerId, int bidAmount) {
+    public BidInfo(int buyerId, Double bidAmount) {
         this.buyerId = buyerId;
         this.bidAmount = bidAmount;
     }
 
     // Getter and setter methods
-    public int getBidAmount() {
+    public Double getBidAmount() {
         return bidAmount;
     }
 
-    public void setBidAmount(int bidAmount) {
+    public void setBidAmount(Double bidAmount) {
         this.bidAmount = bidAmount;
     }
